+++ modules/@apostrophecms/schema/ui/apos/components/AposArrayEditor.vue
@@ -150,7 +150,9 @@ export default {
       return (this.field.schema || []).filter(field => apos.schema.components.fields[field.type]);
     },
     countLabel() {
-      return `${this.next.length} Added`;
+      return this.$t('apostrophe:numberAdded', {
+        count: this.next.length
+      });
     },
     // Here in the array editor we use effectiveMin to factor in the
     // required property because there is no other good place to do that,
@@ -158,14 +160,18 @@ export default {
     // representation of "required".
     minLabel() {
       if (this.effectiveMin) {
-        return `Min: ${this.effectiveMin}`;
+        return this.$t('apostrophe:minUi', {
+          number: this.effectiveMin
+        });
       } else {
         return false;
       }
     },
     maxLabel() {
       if ((typeof this.field.max) === 'number') {
-        return `Max: ${this.field.max}`;
+        return this.$t('apostrophe:maxUi', {
+          number: this.field.max
+        });
       } else {
         return false;
       }